Initial commit of changes to add a SPI secondary mode to avr-hal
Initial commit of changes to add a SPI secondary (or slave) mode to
avr-hal.
This code doesn't work currently.  It is horribly designed.  Some of
the comments are wrong.
It's intended for sharing as a rough starting point.
Areas where there may be issues or room for improvement:
  Specs need to be reviewed for floating vs. pull-up on different SPI pins.
  With better use of Rust features, code could be more DRI.
  Initialization order in the impl_spi macro may be improved upon.

Relevant specs:
  ATmega48A-PA-88A-PA-168A-PA-328-P-DS-DS40002061B.pdf (DS40002061B)
  In particular pages 92 and 170-178